requirements.txt
Flask
flask-cors
pandas
openai
pymysql
cryptography
datetime
mysql-connector-python
cassandra-driver
uuid
